smartmeter/requirements.txt
gurux-common~=1.0.13
gurux-dlms==1.0.197
paho-mqtt~=2.1.0
pycryptodomex~=3.20
pyserial~=3.5
PyYAML~=6.0
lxml~=6.1
beautifulsoup4~=4.14
